Add size option to set number of columns and rows
This commit adds the --size/-s options to Termite to specify the number
of columns and rows for the underlying VTE widget. The accepted format
is {COLS}x{ROWS}, where {COLS} and {ROWS} are base 10 numbers. If the
format is incorrect, an error message is printed and Termite exists with
an error code.

Caveat: from my tests, setting the terminal to size = `CxR` will
actually give you a terminal of size `(C+1)x(R-1)`. I don't know if this
is a font issue or what.